Update to 3.20 release

+++ actioneditorclass.c
@@ -20,7 +20,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  *
- * $Id: actioneditorclass.c,v 1.3 2007/05/08 19:27:07 fab Exp $
+ * $Id: actioneditorclass.c,v 1.4 2025/07/24 01:26:38 geit Exp $
  */
 
 /*
@@ -344,7 +344,7 @@ DEFNEW
 
 		if(!action_node)
 		{
-			return NULL;
+			return ((ULONG) NULL);
 		}
 
 		delete_action_node = TRUE;
@@ -358,7 +358,7 @@ DEFNEW
 		{
 			actionnode_delete(action_node);
 		}
-		return NULL;
+		return ((ULONG) NULL);
 	}
 
 	obj = DoSuperNew(cl, obj,
