+++ columnslistclass.c
@@ -19,7 +19,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  *
- * $Id: columnslistclass.c,v 1.4 2013/10/28 11:00:31 geit Exp $
+ * $Id: columnslistclass.c,v 1.5 2025/07/24 01:26:38 geit Exp $
  */
 
 #include "ambient.h"
@@ -182,7 +182,7 @@ DEFNEW
 
 	if (!obj)
 	{
-		return (NULL);
+		return ((ULONG)NULL);
 	}
 
 	DoMethod(obj, MUIM_Notify, MUIA_Listview_DoubleClick, MUIV_EveryTime,
