+++ ttlistviewclass.c
@@ -18,7 +18,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA
  *
- * $Id: ttlistviewclass.c,v 1.5 2006/04/12 14:01:56 fab Exp $
+ * $Id: ttlistviewclass.c,v 1.6 2025/07/23 23:54:26 geit Exp $
  */
 
 #include "ambient.h"
@@ -45,7 +45,7 @@ DEFNEW
 
 	if (!obj)
 	{
-		return (NULL);
+		return ((ULONG)NULL);
 	}
 
 	return ((ULONG)obj);
